requirements.txt
streamlit
groq
docker
python-dotenv